backend-hackathon/requirements.txt
fastapi
uvicorn
PyPDF2
spacy
google-generativeai
aiofiles
python-multipart
